+++ tests/ai_safety/trustyai_service/constants.py
@@ -1,0 +1,1 @@
+DRIFT_BASE_DATA_PATH: str = "./tests/ai_safety/trustyai_service/drift/model_data"
